refactor: split title_cleanup — pure functions move to livrarr-domain
Pure-function primitives (clean_title, clean_author, title_case,
regex consts, SMALL_WORDS, helpers, 22 tests) relocated from
livrarr-metadata to livrarr-domain. LLM-dependent polish_addtime
stays in livrarr-metadata with re-exports for backward compat.

Prerequisite for the text_norm module (IR v2 CS-001-r1) which
needs clean_title in the domain leaf crate.
