+++ crates/livrarr-domain/Cargo.toml
@@ -13,3 +13,4 @@ trait-variant = { workspace = true }
 tokio = { workspace = true }
 tokio-util = { workspace = true }
 quick-xml = "0.37"
+regex = "1"
